Separation of only obsmon table for time series plots and open individal data bases pr. cycle for other plots.
Operational usage has proven that the observation usage for screening is too large to handle the original design appropiate.

New design works, except for the tab surface diagnostics, which is designed to use the table usage for time series. This must be re-designed do to several queries for the time frame requested.



git-svn-id: https://svn.hirlam.org/trunk/contrib/obsmon2@14628 9085bc96-f90c-0410-a38d-ac8cfabbdae3