Calling obsmon as a command; support for argOpts
The obsmon and obsmon.R files were changed slightly to
make it possible to call obsmon as a command from an
arbitrary directory. To do so, one can, for instance,
put a symlink to obsmon in a directory belonging to the
user's PATH, or, less preferably, putting the whole
obsmon source dir in the user's PATH.

Additionally, the obsmon file was modified so that it passes
any command line arguments it receives to obsmon.R. This may
be useful in the future.

These changes are backwards compatible.
~