Implemented Average Analysis Increment Map.
To to so, a new class of plot, "mapThresholdWithRangeAvgs" was
created. This class behaves much like the previously existing
"mapThreshold", except that it accepts a date range as input and
supports the selection of multiple cycles. The final values to be
plotted will be calculated as averages of the selected dataColumn
over the picked date range and cycles.

A new method, "postProcessQueriedPlotData", was implemented in order
to calculate the averages. The method possess a generic interface of
the form function(plotter, plotData). It calculates the averages, as
discussed above, if called on objects of the mapThresholdWithRangeAvgs
class. Otherwise, it just returns plotData -- thus yielding the same
results as in commits prior to this for all other plot types.

These changes are backwards compatible.
~