Not averaging levels. Better use of aggregate.
The plotData$level column is now added to the "by" list
in the aggregate function, so that averages can also be
obtained as a function of particular levels.

Now passing to aggregate a specific list of plotData columns
to have their values averaged. (the default, used before, was
all). This is more efficient and avoids warnings when attempting
to calculate the mean of non-numeric values (such as statid).
~