Tool to compare R native dep resol. with our own.
R offers tools::package_dependencies, but I've found that the
dependencies it returns are quite ofter not in the proper order, i.e.,
packages appear first which depend on packages appearing later in the
list. I don't know how much this is important for R, but it's certainly
not a good way to resolve dependencies in general.

This tool thus compares R's own routine with the one I've implemented in
the install script. The test shows that they both return the same
dependencies (at least for the packages used thus far in obsmon), but
the routine I've implemented returns the dependencies in the correct
order always.