Automatically determining which pkgs to install.
The only exception is, it seems, the V8 package. V8 is not explicitely used
anywhere in obsmon, but it needs to be installed anyway to prevent the error
"shinyjs: In order to use the `extendShinyjs()` function, you must either use
the `functions` argument, or install the `V8` package with
`install.packages("V8")`".

When removing the previously manually built pkgs list, I realised that
the following packages did not need to have been listed:
  * httpuv: Dependency of shiny
  * mapproj: Suggested dependency of ggplot2,
as they are only used as dependencies of the aforementioned packages
and get installed as part of the dependency resolution.

More worryingly, I could not figure out where the packages below (also
from the previous list) were used:
  * chron
  * jpeg
Not having them installed does not seem to be a problem. Perhaps they
were once needed but are no longer so.