New 'dependencies' func. deps=F in install.packs
The 'dependencies' function was rewriten. It now internally uses
tools::package_dependencies instead of our own previously written
recursive code. Although the previous code itself didn't seem to
be wrong, it seems R recursion was not behaving as expected with
vector arguments in the function. The downside of the tools library
method is, as discussed previously, that it does not return the
dependencies in the correct order. However, this does not seem to
matter as far as R-package installation is concerned. I don't like
this sorting issue, but, well oh well...

Having now a consistent 'dependencies' function (apart from dependency
order), this commit introduces the switch "dependencies=FALSE" to the
install.packages and update.packages functions and all dependencies are
explicitly included in the list of packages to be installed.

Misc changes:
  * The V8 package is now installed automatically as it is listed as a
    "Suggests" dependency of shinyjs
  * update.packages issues warnings if it fails to update a package
    instead of stopping