Saving full dependency list locally to save time.
The full list of packages to be installed is automatically saved in the
utils/build/pkg_dependencies.Rdata. This data is then retrieved when
running the install script, instead of making the checks online. This
saves time during installation. If a new package is added, then the
dependencies for the package will be resolved as usual and the new data
will be added to the Rdata file aforementioned. If one wishes to force
the list to be refreshed, then one shold use the "--refreshdeps" command
line option.

The bahaviour of the "dependencies" routine was changed in order for it
to return character(0) instead of NULL if the dependencies are not
resolved (e.g., if there are no "Depends" section in a package's
DESCRIPTION file). This is to make it easier to implement the dependency
list caching as described in the previous paragraph.