Support to copying pre-compiled pkgs to the R lib
If a tarball named "foo.tgz" or "foo.tar.gz" is found under the
preCompiledPkgsDir directory (see source code for the definition of this
variable), then it is assumed to be a compressed pre-compiled library
and, if it is not yet installed in the system, then it will be unpacked
into the directory defined in the variable "lib" (the same directory any
package would end up in if installed via install.packages).

This has two advantages: First, it will lead to a very quick
installation if pre-compiled pkgs are available for the particulat computer
platform being used. Second, the security risks are reduced by avoiding
the download of source files from the internet (which look fine today,
but may be maliciously changed in the future without us having any
control over what is done).