Install from local srcs if pre-comp. not available
Installing R packages from source files if pre-compiled versios of the
packages are not available. The source files are expected to exist under
the directory ./utils/build/local_R_library/src/. If a needed source is
not found, then the install will fail.

This commit includes all R-package sources needed in the current version of
obsmon.

Automatic download of source files will be implemented in a subsequent
commit, although this should certainly be avoided if one is concerned
about security risks (as discussed in commit 7da8e2f).