Can download missing R package sources from CRAN
This default behaviour of downloading packages was restored in order to
keep backwards compatibility with the current released version of obsmon.
In this commit's version, nonetheless, such downloads will not be necessary,
as all currently needed sources have been included in commit 5d85c8.

The command line option "--abort-if-missing-src" was introduced to force
the install script to abort when srcs are not available. This will most
likely be the way to go when installing obsmon on secure servers.