Binaries now automatically copied to local R lib
The code now passes the "INSTALL_opts="--build"" argument to
install.packages, which causes that routine to create compressed
tarballs of the compiled binaries. These tarballs are then automatically
copied by the install script to the relevant directory under
./utils/build/local_R_library/pre_compiled

This saves time should obsmon needs to be reinstalled by the user (the
code currently only compiles packages for which no binaries are
available). This also saves developers' time, as it is no longer
necessary to manually pack and copy compiled binaries to their due
destinations in the "pre_compiled" directory aforementioned.