Installing multiple pkgs at once is rolled back.
A better approach to copy compiled binaries to the apropriate
utils/build/local_R_library/pre_compiled/ directory was implemented
which does not depend on compiling packages one by one in a for loop.
This is accomplished by using the "finally" clause in a tryCatch
construct, which is used here to call the mvPkgsBinsToRepo function just
created.

Later, this will help parallelisation of the build process.