Back to inst, pkgs 1 by 1. R delets bins otherwise
The code is back installing packages one by one, instead of suporting
parallel build of sources using the Ncpus option of install.packages.
The R routine that installs packages has a clumsy implementation in this
respect and deletes all created binaries upon finishing the parallel
installs.

Unfortunately, whoever wrote the code for R's install.packages routine
decided to put an "unlink(tmpd, recursive = TRUE)" after installation
when it is performed if Ncpus>1. The variable tmpd is defined there as
"tmpd <- file.path(tempdir(), "make_packages")". All tarballs for the
compiled binaries are created in this directory, and there is no way to
change this without changing R. By avoiding the use of Ncpus>1, these
files are not deleted. There is not compelling reason not to loop over
pkgs in this case.

Later I may decide to create tarballs of the libraries myself and turn
parallelisation back on, but I will not do so untill I'm sure there is
no way to do this using routines that are already available.