Add opt to install locally. Lib paths updated.
The commit adds a "--local-install" option to the install script. This
instructs the script to install the libraries under the directory
utils/build/local_R_library/locally_installed_R_library.

This has the advantage of separating obsmon from the rest of the user's
(or the system's) R setup, and is particularly useful when installing
via a packaging system such as RPM, as it allows uninstalling the obsmon
package completely. This also guarantees that installed r libraries will
be available at runtime by all users in the RPM installation case.