Removed source("init.R") from server.R and ui.R
The "obsmon" file already sources init.R, and whatever is in there
should also be accessible in ui.R and server.R as per any reasonable
scoping rule. Not sure if this is the case with Shiny though, so I'll
keep a look and see if this will have to be reverted. My tests so far
indicate this is OK.