Fix crash bug if user cannot read databases
The code now checks whether the user has read access to the needed
databases before attempting to read data from them, thus avoiding sqlite
errors that would lead the code to crash.

If the user does not have read access to the databases listed in the
config file, then a warning will be issued.