Better names for binDirs. EnvVar for binDirs too.
More specific names for the directories that would contain pre-compiled
binaries. These names now contain sysName and sysRelease info, similar
to a previously implemented but then removed behaviour, but with the
difference that the sysRelease number now contains only the major part
of it.

The user now can use the envVar R_LIBS_BIN_REPO_ROOTDIR to point to
directories containing pre-compiled binaries for the R-libs. This is
useful when the user has binaries from other projects and does not want
to have them recompiled.