Looks for deps on remote if new pkg added.
Still looks into the local cran-like repo for dependencies of packages
that we already have locally. Looking on the remote allows more precise
dependency resolution when new packages are added, as these new packages
would not be found in the local repo. This does not mean we are
downloading packages from the remote repo.