Workaround for Shinyjs' "please install V8" bug
Although V8 was correctly identified as a dependency and installed
accordingly, the code would still fail (on Bi@NSC, at least) with a
message from Shinyjs asking for it to be installed. The workaround is to
just explicitely add "library(V8)" somewhere in the code before the
shiny stuff is loaded. Not nice, but, well, oh well...