Caching progress info now shown on the GUI
This info used to be shown in the terminal, which was not very useful
when users were running the code when hosted in a remote server.

Unfortunately I had to implement a scheme using caching status files,
(namely, one exptsCacheProgLogFilePath[[name]] file for each expt name),
as I could not find a way to make server.R see the updates made to the
global variable exptsCachingProgress. To minimise the I/O work created
by this, each of these files is updaded only once the corresponding
caching progress is incremented by 1%. This seems to work well.