New file lib_paths_config.R
The libPaths configuration stuff previously done in the init.R file is now
performed in this new file. This is to prepare for the loading of argparse
when running obsmon as "./obsmon", since it is better to have the command
line arguments parsed before any other type of initialisation occurs.