New option to select only standard levels
The standard levels are considered to be those available in the "obsmon"
table. Levels available only in the "usage" table will not be selected
when "standard" is chosen.

Also, to make it easier to visually tell the levels obtained from the
obsmon and usage tables, the levels obtained from the later are listed
after the levels obtained from the former. The disruption of the sorting
should make the separation clear.