Better backwards-compat to "missing data" bugfix
Better backwards-compatibility strategy to "missing data" bugfix.
Instead of making obsmon thing there was a table called "fromDbTable" in
the chache files and asking users to manually delete the out-of-date
file and restart obsmon, the code now performs the removal of the
problematic files and rebuilds the cache.

This may cause caching times to be longer in the occasion the fix is
applied, but is a more reliable strategy and requires no user
intervention. A warning is issued.