EnvVar OBSMON_MAX_N_EXTRA_PROCESSES: Control nCPUs
Make it possible for users to control the number of extra parallel
processes spawned by obsmon by setting the value of the environment
variable OBSMON_MAX_N_EXTRA_PROCESSES.

The current default is whatever max integer is supported by the machine
running the code. This should be fine because the parallel tasks
currently executed in obsmon are not computationally intensive, but this
value may be changed in the future to something more conservative.

The reason for choosing a high value for this is to prevent the initial
loading of experiments from blocking the rest of the code, as well as to
make sure caching is always non-blocking. It is good however to give
users the option to limit this, and thus the introduction of the new
envVar.