Only show station menu when querying from "usage"
The "backend" obsmon produces output files containing only two tables:
"usage" and "obsmon". Many plots in obsmon are performed with data
queried from the "obsmon" table, but station IDs are only stored in
"usage". Therefore, it does not make much sense to allow users to select
stations when the selected type of plot is such that it will query from
"obsmon".