Can now cancel plot requests
But some investigation still needs to be done to find the cause of a
memory leak.

The memory leak can be observed, e.g., but selecting a very
small/quick kind of plot and performing it multiple times. The execution
time decreases a bit every time, and the amount of memory used by R
increases. It seems to be related to setting "output" inside of the
observeEvent, as commenting out those bits avoids the problem (but
obviously there is no plot shown in this case).