Show spinner when preparing output
Instead of showing a progress barm as it was done previously. The
progress bar was not very informative (it only tracked the progress of
the queries, not of the output itself), and its implementation was more
complicated than the onde for the spinner.

New R-pkg dependencies introduced in this commit:
  shinycssloader
  promises
  later
These correspond to the folling new files:
  shinycssloaders_0.2.0.tar.gz
  promises_1.0.1.tar.gz
  later_0.7.5.tar.gz