oneClickPlots -> quickPlots; Work on multiple quickPlots subplots
Renamed "oneClickPlots" to "quickPlots", as, assuming that the wanted
plot is not the frist on the dropdown menu, one needs to click on the
menu, select the plot name, and only then click on "Plot" -- thus
needing 3 clicks.

Quite a log of work on implementing support to multiple subplots for a
given quickPlot. This is ongoing work; right now only the first of the
generated subplots is sent to the output.

Below, some examples of typical entries in the config.toml file.
If "obs" are omitted, then all obnames that obsmon currently knows
about will be considered. The same for variables (as a function of
the respective obnames, naturally).

[[quickPlots]]
  displayName = "MEPS-mbr000: Number of Observations (CANARI)"
  experiment = "MEPS-mbr000"
  plotType = "Number of Observations"
  database = "ecma_sfc"
  startDate = -30
  [quickPlots.obs]
    synop=["rh2m", "td2m"]
    ship=["snow", "u10m"]

[[quickPlots]]
  displayName = "MEPS-mbr000: Number of Observations (ecma)"
  experiment = "MEPS-mbr000"
  plotType = "Number of Observations"
  database = "ecma"
  startDate = -30
  excludeObs = ["satem"]