Removed spinner from maps and dataTables
Removed these spinners and changed the place where the plot spinner is
inserted because the spinners fail to be removed if two plots are
produced one after another with the same data. Seems to be related to
the shiny bug reported at

"The spinner keeps spinning when rendering a table with the same data"
<https://github.com/andrewsali/shinycssloaders/issues/10>

It is reported as solved there, but it seems to persist here.