Fix shiny memory leak with quickPlots outputs
Unfortunately new variable "quickPlotsMaxNumPltsProduced" had to be
introduced to give an upper limit to the number of plots. This parameter
is configurable via config.toml and has a default value of 500.

The previous strategy seemed better (as it didn't impose an upper
limit), but the code would eventually use up all memory due to this
issue with memory leak in nested reactives in shiny.

This is similar to the issue fixed in <cc3a7bb03aa91c7446>.