Show warning msg on plot if no data or problems
This is similar to how it used to be before commit <1fb94fea21e497>, wth
the differences that (i) The message is put as a text graphic object
instead of via rendering a png file, and (ii) there are two distinct
messages depending on whether problems occured while attempting to query
or whether the query simply returned no data. Having this form of
notification instead of returning a NULL plot is useful for the
quickPlots functionality.

The popup dialog warning that a plot could not be produced has been
removed, as it would be redundant.