Remove parameter quickPlotsMaxNumPltsProduced
And go back to a strategy similar to the one used prior to commit
87921cc7788d to dinamically define the number of (and create the)
needed outputs for the quickPlots. This strategy is based on having
the render* functions inside an observer, but care has been taken to
do some manual garbage collection to avoid (or greatly minimise) any
memory leak (see commits 87921cc7788d and cc3a7bb03aa91c7446).

My tests indicate that this works fine and no leaks seem to be
generated. It seems also that the memory leak issues are related
not just to shiny, but rather to some weird interplay between shiny
and the future package. See commit <d4308d876c2dd>.