quickPlots produced asynchronously
Mind that all plots included in a given quickPlot will be produced in
the same (forked) slave process. For instance, if a quickPlot is to
produce 35 plots, obsmon will NOT create 35 assync processes, but just
one assync process that produces these 35 plots sequentially.

Having this forked process to produce the plots allows the GUI to remain
responsive while the plots are being produced. The idea is to be able to
implement a "cancel quickPlot" button later, analogous to the one
already implemented for the regular plots.