Improve bookkeeping of ongoing caching operations
Also fix bad reactivity of assyncCachingProcs. assyncCachingProcs was a
reactiveValues object and was being used inside an observer without a
proper "isolate" environment, causing the observer in question to fire
every time assyncCachingProcs was changed and causing the UI to become
unresponsive if large amounts of data files were requested to be put in
cache.