v3.0.0-alpha.8: Cache files sent in small batches
So that the files associated with the currently selected Db/DTGs have
priority for caching.

Also streamlined the cache code on server.R, making better use of the
"then" function of the promises package to handle what happens after the
cache finishes. The "cacheFileUpdated" reactive became therefore
obsolete and has been removed.

More descriptive comments added to the bits of the cache code on
server.R.