New "stationSingle" input to select only 1 station
This is going to look and work like the regular "station" input, but
they are never going to appear at the same time. In practice,
stationSingle is a wrapper to station allowing selection of 1 station
only. It makes senso to select multiple stations for some plots, but not
for others, but shiny's updateSelectInput does not allow toggling
multiple selection on and off on demand.