Plots: Add support to zoomming in
The plots are those that are returned as simple ggplot2 objects and use
cartesian coordinates. Namely:
    - First Guess and Analysis Departure
    - All vertical profile plots

For this, the ggplot2 R pkg needed to be updated (there was a bug
in the 2.2.1 version that would cause the plot range to have a
wrong signal when inverting or reversing coordinates). This, in
turn, forced the update of other packages that ggplot2 depend on,
as well as the introduction of new R-pkg dependencies. Summarising:

Updated packages:
    - desc (v1.1.1 --> v1.2.0)
    - ggplot2 (v2.2.1 --> v3.1.0)
    - rlang (v0.2.0 --> v0.3.1)

New packages:
    - askpass
    - callr
    - clipr
    - clisymbols
    - devtools
    - diffobj
    - fontBitstreamVera
    - fontLiberation
    - fontquiver
    - freetypeharfbuzz
    - fs
    - gh
    - git2r
    - ini
    - pkgbuild
    - pkgload
    - processx
    - ps
    - rcmdcheck
    - remotes
    - sessioninfo
    - sys
    - usethis
    - vdiffr
    - whisker
    - xopen