Keep expt and multiPlot choices simple in the GUI
Do not update them dynamically nor separate them between available and
unavailable choices according to the availability, but rather prevent
the user from making plots if the necessary data is unavailable.

Although the previous approach was very quick most of the times, the
code became more complicated than it needs to be, which could make it
difficult for other to maintain the code in the future. Simple is better
if the more complex code does not bring enough advantages.