Fix dtgCacheExpiry: Explicitly set it to Inf
It had inadvertently been left undefined, which only coincidently
produced the desired behaviour.

Setting dtgCacheExpiry to Inf means that the DTGs an instance of
obsmonDatabaseClass has will never be updated. Since new instances are
created upon creation of a new session, the user must refresh the web
page in order to update the experiments' DTGs with additions made to the
data since the current session has been started.