Fix missing error messages when producing plots
Wrap preparePlots and prepareMultiPlots with capture.output in the
shiny app (see new wrapper functions preparePlotsCapturingOutput and
prepareMultiPlotsCapturingOutput), and show eventual outputs produced
during assync plot upon completion of the respective futureCall.

The previous strategy was to sink messages right before the call to
futureCall (sending the output to a file connection), then cancel the
sink right after futureCall, and finally show eventual output diverted
to the file connection upon futureCall completion. This strategy did not
work because the output file connection was not updated with the needed
output upon completion of futureCall.