Add option maxAvgQueriesPerProc to config file
Which controls the max average number of queries a single assync
(future) process is allowed to handle (see perform query routine).

The code had previously fixed this value to 24, but the optimal value
for this is of course system-dependent, so it is better to leave this
for the user to decide.

Default value: Inf
    - This means that all queries will be handled in the same process by
      default, i.e., they will be processed sequentially.