Show help msg in docs if browser can't open pdfs
Some browsers may not have a pdf reader plugin. This will cause the doc
not to be shown, and, rather annoying in some cases, cause the browser
to show a download prompt every time an obsmon tab is open.

To solve this, obsmon now shows a help message when the browser is not
able to open the pdf file.