Append major & minor code version info to cacheDir
To prevent conflict when updating the code. Changes to the way cahing is
written and read should imply updates in the major or at least minor
fields of the code version. The patch number has therefore been ignored
in this particular case, to prevent the creation of too many cache
directories as obsmon is updated.