Get obs attrs to be cached directly form cache files
Get, from the sql cache files themselves:
    1) The observation types supported for caching
    2) The attributes to be cached for each observation type

This had up to now been defined via a list named "obsKeyAttributes" in
the cache_routines.R file, which meant that every time a new observation
type was added, one would have to remember to add both a new table in
the template cache file and a new corresponding entry in this
obsKeyAttributes list. Now, it suffices to add a new table in the
template cache file.