Reset DTG cache if srcDB mtime > cache entry ctime
Reset the cache entries related to a given DTG if the associated source
DB file appears to have been modified after the original cache entry was
created. This is useful when people are testing stuff and modifying
experiment files. If this is not done, the cached attributes may not be
consistent with the actual information in the source files.