Protect essential deps of suggest-type deps
Prevent essential dependencies of suggest-type dependencies from being
removed from the list of libraries that need to be installed. This could
happen, for instance, if such libraries were listed in the array named
"excludeFromSuggests".

This is necessary because otherwise some suggests-type libs not listed
as excludeFromSuggests may fail to build due to their dependencies being
listed in excludeFromSuggests.