Make map plots interactive (purely plotly)
And keep the original ggplot2 routines for when it is requested that
plots be non-interactive.

The plotly package needed to be updated to solve a bug with setting
marker colours (the bug would cause most points to vanish in this case).
Some ner R-libraries needed to be added/updated as a result:

R-libraries updated:
  plotly: v4.8 --> v4.9
  jsonlite: v1.5 -- > v1.6

R-libraries added:
  debugme v1.1.0
  parsedate v1.2.0
  pingr v2.0.0
  rematch v1.0.1
  rematch2 v2.1.0
  shinytest v1.3.1
  showimage v1.0.0
  webdriver v1.0.5