Install: Add R version to pre-compiled bins paths
Add info on major.minor version of R to the paths where pre-compiled
binaries are expected to be found, so that the install script
effectively requires that build and install major.minor versions of R
match. This helps preventing issues with incompatibility of R libraries
compiled with R versions that do not match the one used when actually
loading the libraries.