Can make plots for new obs without registering
New obs of usual obtypes (surface, upper air, satellite, etc) are
now supported in some plots without the need registering to register
them in the observation_definitions.R. To do this, some changes in the
caching routines were made.

Not all plots work yet with this (obs usage, for instance), so registering
the new observation type on observation_definitions.R is still the way
to go for a definitive solution.

Close this for the moment as netatmo obs are needed and this is already
covered.