Change spinner to gif. Use it while making plot.
The previously used spinners have been replaced by a gif image. The same
gif image is used in an empty plot shown while the real plot is being
produced. We did this because the previously used spinner scheme didn't
allow the spinners to be shown while the plots were being produced
asynchronously, as we used to set the chart to NULL while waiting for
completion.