Default to not having .installer_pkg_versions.txt
Default to not having a ".installer_pkg_versions.txt" file.
There is already a way to lock the versions, namely the file
".installer_local_pkg_repo/src/contrib/PACKAGES".

Leave the option to have the ".installer_pkg_versions.txt" in case one
needs to recreate the local repo and/or lock some packages to specific
versions. It is easier for users to edit the .installer_pkg_versions.txt
file than the PACKAGES one.